instanced component system started?

+++ scripts/delete_cmake_files.sh
@@ -4,6 +4,8 @@
 SELF_SCRIPT="$(basename "$0")"
 PROTECTED_SCRIPT_1="build_and_run.sh"
 PROTECTED_SCRIPT_2="clean_build_and_run.sh"
+PROTECTED_SCRIPT_3="debug_build.sh"
+PROTECTED_SCRIPT_4="compile_commands.json" # For vscode
 
 # Loop through all items (including hidden ones) in the current directory
 for item in * .*; do
@@ -11,7 +13,7 @@ for item in * .*; do
   [[ "$item" == "." || "$item" == ".." ]] && continue
 
   # Skip the protected scripts
-  [[ "$item" == "$SELF_SCRIPT" || "$item" == "$PROTECTED_SCRIPT_1" || "$item" == "$PROTECTED_SCRIPT_2" ]] && continue
+  [[ "$item" == "$SELF_SCRIPT" || "$item" == "$PROTECTED_SCRIPT_1" || "$item" == "$PROTECTED_SCRIPT_2" || "$item" == "$PROTECTED_SCRIPT_3" || "$item" == "$PROTECTED_SCRIPT_4" ]] && continue
 
   # Delete the item
   rm -rf -- "$item"
